+++ testdata/statement/remove.go.12.go
@@ -41,7 +41,6 @@ func foo() int {
 		n--
 	default:
 		_ = n
-
 		fmt.Println(n)
 		func() {}()
 	}